# Use number instead of JSBI.BigInt for smaller integers