--- a/src/dom/Integer.ts
+++ b/src/dom/Integer.ts
@@ -96,18 +96,10 @@ export class Integer extends Value(Number, IonTypes.INT, _fromJsConstructor) {
   ): boolean {
     let isSupportedType: boolean = false;
     let valueToCompare: any = null;
-    if (other instanceof Integer) {
-      isSupportedType = true;
-      if (this._bigIntValue == null && other._bigIntValue == null) {
-        valueToCompare = other.numberValue();
-      } else {
-        // One of them is a JSBI
-        valueToCompare = other.bigIntValue();
-      }
-    }
 
     if (options.onlyCompareIon) {
       // `compareOnlyIon` requires that the provided value be an ion.dom.Integer instance.
+      isSupportedType = true;
       if (this._bigIntValue == null && other._bigIntValue == null) {
         valueToCompare = other.numberValue();
       } else {