--- a/src/dom/Integer.ts
+++ b/src/dom/Integer.ts
@@ -99,12 +99,14 @@ export class Integer extends Value(Number, IonTypes.INT, _fromJsConstructor) {
 
     if (options.onlyCompareIon) {
       // `compareOnlyIon` requires that the provided value be an ion.dom.Integer instance.
-      isSupportedType = true;
-      if (this._bigIntValue == null && other._bigIntValue == null) {
-        valueToCompare = other.numberValue();
-      } else {
-        // One of them is a JSBI
-        valueToCompare = other.bigIntValue();
+      if (other instanceof Integer) {
+        isSupportedType = true;
+        if (this._bigIntValue == null && other._bigIntValue == null) {
+          valueToCompare = other.numberValue();
+        } else {
+          // One of them is a JSBI
+          valueToCompare = other.bigIntValue();
+        }
       }
     } else {
       // We will consider other Integer-ish types