--- a/src/events/IonEvent.ts
+++ b/src/events/IonEvent.ts
@@ -453,10 +453,16 @@ class IonIntEvent extends AbstractIonEvent {
   valueCompare(expected: IonEvent): ComparisonResult {
     if (expected instanceof IonIntEvent) {
       // convert both values to `JSBI.BigInt` first and then compare for precision.
-      let actualValue = this.ionValue instanceof JSBI ? this.ionValue : JSBI.BigInt(this.ionValue);
-      let expectedValue = expected.ionValue instanceof JSBI ? expected.ionValue : JSBI.BigInt(expected.ionValue);
+      let actualValue =
+        this.ionValue instanceof JSBI
+          ? this.ionValue
+          : JSBI.BigInt(this.ionValue);
+      let expectedValue =
+        expected.ionValue instanceof JSBI
+          ? expected.ionValue
+          : JSBI.BigInt(expected.ionValue);
 
-      if(JSBI.equal(actualValue, expectedValue)) {
+      if (JSBI.equal(actualValue, expectedValue)) {
         return new ComparisonResult(ComparisonResultType.EQUAL);
       }
     }