*Issue #663*

*Description of changes:*
This PR works on a performance change using `number` instead of `JSBI.BigInt` for smaller integers.

*Changes:*
1. Change IonBinaryParser logic to use `number` for integers smaller than 6 bytes. Otherwise, use `JSBI.BigInt`. 
2. Modified dom.Value#Integers to use `JSBI.equal` for comparison. Hence converting both comparison values to `JSBI.BigInt` first for precision.
3. Modified the `IonEvent#valueCompare` for `Integer` Event to convert both comparison values to `JSBI.BigInt` first and then compare.

*Test:*
Verified performance using [benchmarkjs](https://benchmarkjs.com) with large binary ion file as input.

*Test Results:*
| Input file type  | Time(before enhancement) | Time (after enhancement) | % decrease in Time
|----------|:-------------:|------:|-----:|
| Binary - big integers | 12.666 | 12.820 | -1.215 
| Binary - small integers | 10.356 | 1.000 | 90.342